Graph updated
The graph functionality has been updated.
* horizontal lines are now constricted to the graph axis by introducing
a new parent div. Use has been made of the pointer-events css property,
which means IE is a problem…
* horizontal lines now have the required states (dashed, solid, grey).
When flipping buttons the correct states are now visible through the
horizontal lines.
* graph styling is adjusted to more closely resemble the old design
* fixed a small bug in bet_events.js that statically called the
buttonStates template with 'sell' as a variable instead of 'buy' or
'sell' depending on the button that was clicked.