login/logout menu item
Added frontend functionality for logging in/out. I named the div
"status" . Includes the html (in the template statusStates.ejs) for the
status "loggedIn" or "loggedOut". In menu_new.js some functions are
changed to allow sliding in/out of the menu, changing the arrow
direction and removing/displaying the watermark in the "username" and
"password" input fields. The image "menu.png" has been changed to also
include the sprites necessary for this layout.
Some other parts have been updated (charts.js, and a few chart-related
bet.php alterations) for testing with the zooming functionality of the
chart.