chart update
Extended currentY variable to be an array (for each outcome a
'currentY' value that should be set by the Pusher message receiver
handler.
The frontend should receive match data before it can be initialized.
That should for now be:
(1) the latest stock values for each outcome (the currentY array values
must be set). The graph will then extrapolate data for the last 5
minutes, which is currently the unadjustable zoom level.
(2) the game data to be rendered (a variable containing the following
match data, for example:

var dataOutcomes = {
  title : "titel",
  outcomes : ['Nederland', 'Italië', 'Gelijkspel']
};

A split is also made within the function chartData(). Now the var
"demo" can be set to true so that random stock developments will be
generated.